move storage level test StorageLevelSpec class to avoid unrelated errors in other tests
